moves base class to its own file
