+++ cg/store/api/core.py
@@ -12,28 +12,8 @@ from .trends import TrendsHandler
 LOG = logging.getLogger(__name__)
 
 
-class BaseHandler:
-
-    User = models.User
-    Customer = models.Customer
-    CustomerGroup = models.CustomerGroup
-    Sample = models.Sample
-    Family = models.Family
-    FamilySample = models.FamilySample
-    Flowcell = models.Flowcell
-    Analysis = models.Analysis
-    Application = models.Application
-    ApplicationVersion = models.ApplicationVersion
-    Panel = models.Panel
-    Pool = models.Pool
-    Delivery = models.Delivery
-    Invoice = models.Invoice
-    MicrobialSample = models.MicrobialSample
-    MicrobialOrder = models.MicrobialOrder
-    Organism = models.Organism
-
-
-class CoreHandler(BaseHandler, AddHandler, FindHandler, StatusHandler, TrendsHandler):
+class CoreHandler(AddHandler, FindHandler, StatusHandler, TrendsHandler):
+    """Aggregating class for the store api handlers"""
     pass
