+++ cg/store/api/trends.py
@@ -1,19 +1,20 @@
 import datetime as dt
 from itertools import groupby
-from dateutil.relativedelta import relativedelta
-from sqlalchemy import func, text
+from sqlalchemy import func
 
 from cg.store import models
+from cg.store.api.base import BaseHandler
 
 MONTHS = {1: 'January', 2: 'February', 3: 'March', 4: 'April', 5: 'May', 6: 'June',
           7: 'July', 8: 'August', 9: 'September', 10: 'October', 11: 'November', 12: 'December'}
 
 
-class TrendsHandler:
+class TrendsHandler(BaseHandler):
+    """Deprecated in favour of Vogue"""
 
     @staticmethod
     def get_last_day_of_previous_year(year: int):
-        return TrendsHandler.get_last_day_of_year(year-1)
+        return TrendsHandler.get_last_day_of_year(year - 1)
 
     @staticmethod
     def get_last_day_of_year(year):
@@ -35,18 +36,18 @@ class TrendsHandler:
                 func.month(models.Sample.received_at).label('month_no'),
                 func.count(models.Sample.id).label('count'),
             )
-                .join(models.Sample.customer)
-                .filter(models.Sample.received_at > self.get_from_date(year),
-                        models.Sample.received_at < self.get_until_date(year))
-                .group_by('priority','month_no')
+            .join(models.Sample.customer)
+            .filter(models.Sample.received_at > self.get_from_date(year),
+                    models.Sample.received_at < self.get_until_date(year))
+            .group_by('priority', 'month_no')
         )
 
         for cust_priority, results in groupby(query, key=lambda result: result.priority):
             counts = {MONTHS[result.month_no]: result.count for result in results}
             data = {
                 'name': cust_priority,
-                'results': {month : counts.get(month) or None
-                 for month in MONTHS.values()},
+                'results': {month: counts.get(month) or None
+                            for month in MONTHS.values()},
             }
             yield data
 
@@ -76,11 +77,10 @@ class TrendsHandler:
             counts = {MONTHS[result.month_no]: result.count for result in results}
             data = {
                 'name': application_cathegory,
-                'results': {month : counts.get(month) or None
-                 for month in MONTHS.values()},
+                'results': {month: counts.get(month) or None
+                            for month in MONTHS.values()},
             }
             yield data
-    
 
     def received_to_delivered(self, year):
         """Calculate averages from received to delivered."""
@@ -112,8 +112,8 @@ class TrendsHandler:
                         for result in results}
             yield {
                 'name': category,
-                'results': {month : averages.get(month) or None
-                 for month in MONTHS.values()}
+                'results': {month: averages.get(month) or None
+                            for month in MONTHS.values()}
             }
 
     def received_to_prepped(self, year):
@@ -145,8 +145,8 @@ class TrendsHandler:
                         for result in results}
             yield {
                 'name': category,
-                'results': {month : averages.get(month) or None
-                 for month in MONTHS.values()}
+                'results': {month: averages.get(month) or None
+                            for month in MONTHS.values()}
             }
 
     def prepped_to_sequenced(self, year):
@@ -179,8 +179,8 @@ class TrendsHandler:
                         for result in results}
             yield {
                 'name': category,
-                'results': {month : averages.get(month) or None
-                 for month in MONTHS.values()}
+                'results': {month: averages.get(month) or None
+                            for month in MONTHS.values()}
             }
 
     def sequenced_to_delivered(self, year):
@@ -214,8 +214,8 @@ class TrendsHandler:
                         for result in results}
             yield {
                 'name': category,
-                'results': {month : averages.get(month) or None
-                 for month in MONTHS.values()}
+                'results': {month: averages.get(month) or None
+                            for month in MONTHS.values()}
             }
 
     def delivered_to_invoiced(self, year):
@@ -254,6 +254,6 @@ class TrendsHandler:
                         for result in results}
             yield {
                 'name': category,
-                'results': {month : averages.get(month) or None
-                 for month in MONTHS.values()}
+                'results': {month: averages.get(month) or None
+                            for month in MONTHS.values()}
             }
